Pull App providers into component
